refactor!: split playa.page into three modules
BREAKING CHANGE: content objects are in playa.content now
