+++ CHANGELOG.md
@@ -10,6 +10,8 @@
   `playa.pdftypes`
 - BREAKING: `PathObject` no longer contains "subpaths", it is safe to
   recursively descend it now
+- BREAKING: Content objects moved to `playa.content` and interpreter
+  to `playa.interp`
 
 ## PLAYA 0.4.2: 2025-04-26
 - Correct `fontsize` and `scaling` in text state
